move all integration tests from common into root
