+++ apps/crn-server/test/integration/events.integration-test.ts
@@ -1,9 +1,9 @@
 import supertest from 'supertest';
 import { Express } from 'express';
 
-import { PAGE_SIZE } from '../../../scripts/export-entity';
-import { AppHelper } from '../helpers/app';
-import { FixtureFactory, getUserFixture, UserFixture } from '../fixtures';
+import { PAGE_SIZE } from '../../scripts/export-entity';
+import { AppHelper } from './helpers/app';
+import { FixtureFactory, getUserFixture, UserFixture } from './fixtures';
 
 jest.setTimeout(120000);
