+++ jest-integration.config.js
@@ -6,8 +6,5 @@ const {
 module.exports = {
   ...base,
   collectCoverageFrom: ['<rootDir>/apps/crn-server/src/**/*.{js,jsx,ts,tsx}'],
-  testMatch: [
-    '<rootDir>/**/common/*.integration-test.{js,jsx,ts,tsx}',
-    `<rootDir>/**/contentful/*.integration-test.{js,jsx,ts,tsx}`,
-  ],
+  testMatch: ['<rootDir>/**/*.integration-test.{js,jsx,ts,tsx}'],
 };
